pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edanalyzer"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
description = "One line description of your module"
dependencies = [
    "gemmi",
    "loguru",
    "SQLAlchemy",
    "pydantic",
    "networkx",
    "numpy",
    "rich",
    "fire",
    "numpy",
    "pandas",
    "joblib",
    "torch",
    "torchvision",
    "scipy",
    "pyyaml"
] # Add project dependencies here, e.g. ["click", "numpy"]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.rst"
requires-python = ">=3.10"